% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.mars.R
\name{anova.mars}
\alias{anova.mars}
\title{ANOVA for MARS Model Objects}
\usage{
\method{anova}{mars}(object)
}
\arguments{
\item{object}{An object of class \code{'mars'}, typically resulting from a call to \code{\link[mars]{mars()}}.}
}
\value{
Returns a comprehensive ANOVA table for the MARS model, including statistics for each
basis function. The output is analogous to the ANOVA output for linear models but tailored
to the specifics of the MARS model framework.
}
\description{
This function extends the generic \code{anova} method for objects of class \code{'mars'}.
It provides detailed ANOVA statistics for the fitted MARS model, similar to the traditional
ANOVA output for linear models (\code{lm}). Additionally, it includes specific details about the
basis functions used in the model, enhancing understanding of the model structure and fit.
}
\examples{
# Fit a MARS model to the concrete dataset
mm <- mars(ConcreteCompressiveStrength ~ ., data = concrete, control = mars.control(Mmax = 10, d = 3, trace = FALSE))
# View the ANOVA summary for the fitted model
output <- anova(mm)

}
\references{
\enumerate{
\item Friedman, J. H. (1991). "Multivariate Adaptive Regression Splines." The Annals of Statistics, 19(1), pp. 1-67.
Available online: \href{https://www.jstor.org/stable/2241837}{JSTOR}
\item STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}
}
}
