% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{bwd_stepwise}
\alias{bwd_stepwise}
\title{Backward stepwise function based on Algorithm 3 by Freifdman}
\usage{
bwd_stepwise(fwd, control)
}
\arguments{
\item{fwd}{output of forward stepwise function, it is a list of vector of
response, Basis function matrix and Bfuncs which recorded the information of
basis functions.}

\item{control}{an object of class 'mars.control': a list of Mmax, d, and trace,
see helper file mars.control() for more information.}
}
\value{
a list containing the following information:
\itemize{
\item{y:}{ the column vector of response variable.}
\item{B:}{ a matrix that contains information of basis functions reduced from
the B in forward selection to avoid overfitting. Each column of basis functions are
products of hinge functions with information of sign, number of predictor and
split point given in Bfuncs in order.}
\item{Bfuncs:}{ a list containing information on how each basis function
is created, each element of Bfuncs correspond to the hinge functions that are
used to create each column of B. Inside each element of Bfuncs, there could be
null, 1 or multiples rows. Each row represents information of each hinge function.}
}
}
\description{
Base on the output of forward selection, perform backward selection using
basis functions to reduce over-fitting.
}
\examples{
fwdtest <- fwd_stepwise(y=concrete$ConcreteCompressiveStrength,x=concrete[,-1],
control = mars.control(8,3,FALSE))

bwdtest <- bwd_stepwise(fwdtest,control = mars.control(8,3,FALSE))

}
\references{
i)

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837

ii)
STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}
}
