% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-concrete.R
\docType{data}
\name{concrete}
\alias{concrete}
\title{Concrete Compressive Strength Dataset for Package}
\format{
A data frame with 100 observations and 9 features:
\describe{
\item{ConcreteCompressiveStrength}{Numeric: Compressive strength of concrete (MPa), response variable.}
\item{Cement}{Numeric: Quantity of cement (kg in a cubic meter mixture), explanatory variable.}
\item{BlastFurnaceSlag}{Numeric: Quantity of blast furnace slag (kg in a cubic meter mixture), explanatory variable.}
\item{FlyAsh}{Numeric: Quantity of fly ash (kg in a cubic meter mixture), explanatory variable.}
\item{Water}{Numeric: Quantity of water (kg in a cubic meter mixture), explanatory variable.}
\item{Superplasticizer}{Numeric: Quantity of superplasticizer (kg in a cubic meter mixture), explanatory variable.}
\item{CoarseAggregate}{Numeric: Quantity of coarse aggregate (kg in a cubic meter mixture), explanatory variable.}
\item{FineAggregate}{Numeric: Quantity of fine aggregate (kg in a cubic meter mixture), explanatory variable.}
\item{Age}{Numeric: Age of the concrete (days), explanatory variable.}
}
}
\usage{
data(concrete)
}
\description{
This dataset comprises 100 observations used to predict the concrete compressive strength,
a critical measure in civil engineering. It includes eight explanatory variables
representing different components and characteristics of the concrete mix, along with the
response variable representing the concrete's compressive strength measured in MPa (Mega Pascals).
}
\details{
The dataset is particularly useful for regression analysis, enabling predictions
of concrete strength based on its composition and age. Such analyses are valuable
in construction and material science for ensuring the quality and durability of concrete structures.
}
\references{
This dataset is a shortened version (first 100 observations) of the original dataset available from the UCI Machine Learning Repository:

\href{https://archive.ics.uci.edu/ml/datasets/Concrete+Compressive+Strength}{Concrete Compressive Strength Dataset}

It is utilized for illustrative purposes in the package.
}
\keyword{datasets}
