% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{h}
\alias{h}
\title{Apply Hinge Function to Predictor Variable}
\usage{
h(x, s, t)
}
\arguments{
\item{x}{Numeric vector representing the predictor variable.}

\item{s}{Numeric scalar indicating the direction of the hinge function: +1 for values greater than the split point and -1 for values less than the split point.}

\item{t}{Numeric scalar specifying the split point (knot) where the hinge function effect begins.}
}
\value{
Returns a numeric vector where each element is adjusted according to the hinge function: elements greater (or less) than the split point are modified based on the sign \code{s}; elements not meeting the condition are set to zero.
}
\description{
The hinge function is a fundamental component in constructing basis functions for Multivariate Adaptive Regression Splines (MARS). It modifies the predictor variable values based on a specified split point and direction (sign). This function facilitates non-linear mapping of the input variable, enabling piecewise linear modeling in MARS.
}
\examples{
# Apply hinge function to the 'Cement' predictor in the concrete dataset
# with a split point at 300 and positive direction
adjusted_values <- h(x = concrete$Cement, s = 1, t = 300)

}
\references{
Friedman, J. H. (1991). "Multivariate Adaptive Regression Splines." The Annals of Statistics, 19(1), pp. 1-67. \href{https://www.jstor.org/stable/2241837}{DOI: https://doi.org/10.1214/aos/1176347963}

STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}.
}
