% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{init_B}
\alias{init_B}
\title{Initialize Basis Function Matrix for MARS Model}
\usage{
init_B(N, Mmax)
}
\arguments{
\item{N}{The total number of observations in the dataset.}

\item{Mmax}{The maximum number of basis functions (excluding the intercept)
to be considered during the model fitting process. This value is dictated by
the 'Mmax' parameter in a \code{mars.control} object.}
}
\value{
Returns a matrix where the first column is filled with ones to
represent the intercept, and the remaining columns are initialized with NA
values, awaiting the assignment of calculated basis functions. The dimensions
of the matrix are N rows by \code{Mmax} + 1 columns.
}
\description{
Initializes a matrix intended to store basis functions during the forward
stepwise selection process of a MARS model fitting. The matrix is pre-filled
with a column of ones to represent the intercept and NA values for placeholders
of potential basis functions to be determined.
}
\examples{
# Initializing a basis function matrix for 100 observations with up to 10 basis functions
Bmatrix <- init_B(N = 100, Mmax = 10)

}
\references{
Friedman, J. H. (1991). "Multivariate Adaptive Regression Splines." The Annals of Statistics,
19(1), pp. 1-67. \href{https://www.jstor.org/stable/2241837}{Link to the article}.

STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}
}
