% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{LOF}
\alias{LOF}
\title{Calculate Lack of Fit (LOF) Using Generalized Cross-Validation (GCV)}
\usage{
LOF(form, data, control)
}
\arguments{
\item{form}{A \code{formula} object for the linear model, specifying the response variable
and predictors. It must include \code{-1} to explicitly exclude the intercept, aligning with
MARS model conventions.}

\item{data}{A \code{data.frame} containing both the response and predictor variables as
specified in \code{form}.}

\item{control}{A \code{mars.control} object containing tuning parameters for the MARS model
fitting process: \code{Mmax} (maximum number of basis functions), \code{d} (penalty parameter for
model complexity), and \code{trace} (boolean flag to enable tracing of the fitting process).}
}
\value{
The function returns the LOF metric calculated using the GCV criterion for the
specified model and data.
}
\description{
Computes the Lack of Fit (LOF) metric, applying the Generalized Cross-Validation (GCV)
criterion for a given linear model. The GCV criterion is defined as:
\deqn{RSS \times N / (N - \tilde{C}(M))^2}, where \eqn{RSS} is the residual sum of
squares, \eqn{N} is the number of observations, and \eqn{\tilde{C}(M)} is the effective
degrees of freedom, calculated as the sum of the diagonal elements of the hat matrix
(hat-values) plus the product of \eqn{d} and the number of model coefficients excluding
the intercept.
}
\examples{
lof_value <- LOF(
  form = ConcreteCompressiveStrength ~ . - 1,
  data = concrete,
  control = mars.control()
)

}
\references{
Friedman, J. H. (1991). "Multivariate Adaptive Regression Splines." The Annals of
Statistics, 19(1), pp. 1-67. \href{https://www.jstor.org/stable/2241837}{DOI: https://doi.org/10.1214/aos/1176347963}

For further examples and insights into MARS model fitting, refer to the STAT 360
STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}.
}
