% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{make_B}
\alias{make_B}
\title{Construct Basis Function Matrix for 'mars' Model}
\usage{
make_B(X, Bfuncs)
}
\arguments{
\item{X}{The model matrix containing predictors only. This matrix should correspond to
the new data for which predictions are intended, or the original data used to
fit the model if predictions on the training data are desired.}

\item{Bfuncs}{A list detailing the basis functions as derived from a 'mars' model object.
This list encapsulates the model's structural components, including
interactions and transformations applied to the original predictors.}
}
\value{
Returns a matrix where each column represents a basis function derived from
the input predictors, ready to be used for predictions with a 'mars' model's coefficients.
}
\description{
Utilizes the basis function information (\code{Bfuncs}) derived from a fitted 'mars' model
and the predictor data to construct the basis function matrix essential for making predictions.
}
\examples{
# Assuming tmars is a fitted mars model:
tr <- terms(tmars$formula)
mf <- model.frame(tr, concrete, na.action = na.exclude)
X <- model.matrix(tr, mf)[, -1, drop = FALSE]  # Exclude the intercept
B <- make_B(X, tmars$Bfuncs)

}
\references{
Friedman, J. H. (1991). "Multivariate Adaptive Regression Splines." The Annals of Statistics,
Vol. 19, No. 1, pp. 1-67. Available at \href{https://www.jstor.org/stable/2241837}{JSTOR}.

Additional examples and materials on MARS and the use of \code{make_B} can be found in the
STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}.
}
