% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.control}
\alias{mars.control}
\title{Construct 'mars.control' Object for MARS Model Fitting}
\usage{
mars.control(Mmax = 2, d = 3, trace = FALSE)
}
\arguments{
\item{Mmax}{The maximum number of basis functions to include in the model, excluding
the intercept. Must be an even integer greater than or equal to 2. Default is 2.}

\item{d}{Degree of penalty in the GCV criterion, controlling the trade-off between
model fit and complexity. Jerome H. Friedman suggests d = 3 as a generally effective
value. Default is 3.}

\item{trace}{Logical flag to enable verbose output during the model fitting process,
particularly useful for tracking the addition of candidate basis functions and their
impact on the LOF (Lack of Fit) measured by GCV. Default is FALSE.}
}
\value{
Returns a \code{mars.control} object, a list containing the specified settings
for MARS model fitting.
}
\description{
Constructs a \code{mars.control} object specifying the configuration parameters for
Multivariate Adaptive Regression Splines (MARS) model fitting. This includes settings
for the maximum number of basis functions, the parameter for Generalized Cross-Validation
(GCV) calculation, and an option to enable tracing of the model fitting process.
}
\examples{
# Default settings
mc_default <- mars.control()

# Customized settings
mc_custom <- mars.control(Mmax = 10, d = 3, trace = TRUE)

}
\references{
Friedman, J. H. (1991). "Multivariate Adaptive Regression Splines." The Annals of Statistics,
Vol. 19, No. 1, pp. 1-67. \href{https://www.jstor.org/stable/2241837}{Link}

For more information and examples on MARS model fitting, see the STAT 360 course materials
STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}.
}
