% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Fit Multivariate Adaptive Regression Splines (MARS) Model}
\usage{
mars(formula, data, control = mars.control())
}
\arguments{
\item{formula}{A \code{formula} object, similar to those used in \code{lm()}, specifying the model to be fitted. The formula must include the response variable and one or more predictors.}

\item{data}{A \code{data.frame} containing the dataset over which the model is to be fitted, including both the response and the predictor variables.}

\item{control}{A \code{mars.control} object specifying the tuning parameters for the MARS algorithm: \code{Mmax} (maximum number of basis functions), \code{d} (degree of interaction), and \code{trace} (logical flag for step-by-step tracing of model selection). Defaults are set based on Friedman's recommendations. For detailed configuration, refer to \code{?mars.control}.}
}
\value{
Returns an object of class \code{mars}, containing comprehensive details of the fitted model, analogous to \code{lm} objects but enriched with specific MARS attributes:
\itemize{
\item{call}{The original function call.}
\item{formula}{The model formula used.}
\item{y}{Response variable vector.}
\item{B}{Matrix of basis functions selected during model fitting.}
\item{Bfuncs}{Details on the construction of each basis function.}
\item{x_names}{Names of predictor variables.}
\item{coefficients}{Model coefficients.}
\item{residuals}{Residuals from the fitted model.}
\item{effects}{Orthogonal effects.}
\item{rank}{Rank of the model.}
\item{fitted.values}{Fitted values.}
\item{assign}{Assignment of coefficients.}
\item{qr}{QR decomposition of the model matrix.}
\item{df.residual}{Degrees of freedom for residuals.}
\item{xlevels}{Levels of factors used in fitting.}
\item{terms}{Terms object from the model fitting.}
\item{model}{The model frame used.}
}
}
\description{
Implements Friedman's Multivariate Adaptive Regression Splines (MARS) to model complex relationships between a response variable and multiple predictors. MARS models are capable of capturing non-linearities and interactions between variables through piecewise linear or polynomial functions.
}
\examples{
# Fitting a MARS model to predict concrete compressive strength
mm <- mars(ConcreteCompressiveStrength ~ ., data = concrete)

}
\references{
Friedman, J. H. (1991). "Multivariate Adaptive Regression Splines." The Annals of Statistics, 19(1), pp. 1-67. \href{https://www.jstor.org/stable/2241837}{Link}

STAT 360 GitHub Repository by Becky Lin: \href{https://github.com/Becky07/STAT360}{STAT360}

R Documentation on Linear Models (\code{lm}): \href{https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/lm}{lm in R}

R Documentation on Model Effects: \href{https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/effects}{effects in R}

R Documentation on Model Terms: \href{https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/terms}{terms in R}
}
