% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{new_mars.control}
\alias{new_mars.control}
\title{Constructor for 'mars.control' objects}
\usage{
new_mars.control(control)
}
\arguments{
\item{control}{A list specifying the values for Mmax, d, and trace.
Mmax controls the maximum number of basis functions, including both
main effects and interactions. The d parameter is used in the calculation
of the generalized cross-validation (GCV) score, penalizing model complexity.
Setting trace to TRUE enables printing of detailed output during the model
fitting process, which can help in understanding the selection of basis functions.}
}
\value{
Returns an object of class 'mars.control', ready to be used in MARS model fitting.
}
\description{
Creates a 'mars.control' object containing parameters that dictate how the MARS model
will be fitted. This function allows users to specify their preferences for the
maximum number of basis functions (Mmax), the penalty parameter (d) for model complexity
in the GCV criterion, and whether to trace the model fitting process (trace).
}
\examples{
# Creating a mars.control object with specific parameters
mc <- new_mars.control(list(Mmax=10, d=3, trace=TRUE))

}
\references{
i) Paper: Multivariate Adaptive Regression Splines
Author(s): Jerome H. Friedman
Source: The Annals of Statistics, Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67
Published by: Institute of Mathematical Statistics Stable
URL: https://www.jstor.org/stable/2241837

ii) STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}.

The constructor simplifies the process of setting up MARS model parameters, while the
validator ensures the provided values are appropriate for model fitting, improving
usability and robustness of the MARS modelling process.
}
