% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mars.R
\name{plot.mars}
\alias{plot.mars}
\title{Visualize Model Diagnostics for mars Objects}
\usage{
\method{plot}{mars}(object)
}
\arguments{
\item{object}{An object of class \code{'mars'}, typically the result of fitting a MARS model
using the \code{\link{mars}} function.}
}
\value{
Invisibly returns \code{NULL}. The function is called for its side effect
of producing plots.
}
\description{
Generates diagnostic plots for evaluating the fit of a MARS model. This method
creates three standard plots: Residuals vs. Fitted values, Histogram of Residuals,
and a Normal Q-Q plot of Residuals, which are crucial for assessing the model's
assumptions and performance.
}
\examples{
# Assuming 'mm' is a fitted 'mars' model object
plot(mm)

}
\references{
Friedman, J. H. (1991). "Multivariate Adaptive Regression Splines." The Annals of Statistics,
Vol. 19, No. 1, pp. 1-67. \href{https://www.jstor.org/stable/2241837}{Link}

Additional examples and materials on MARS can be found in the STAT 360 GitHub repository
STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}.
}
\seealso{
\code{\link{mars}} for model fitting, \code{\link{residuals}} for extracting residuals.
}
