% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{predict.mars}
\alias{predict.mars}
\title{Prediction for mars Model Objects}
\usage{
\method{predict}{mars}(object, newdata)
}
\arguments{
\item{object}{A \code{mars} object resulting from fitting a MARS model using the \code{\link{mars}} function.}

\item{newdata}{An optional data frame containing the predictors for which predictions are desired.
If omitted, predictions are made for the original data used to fit the model.}
}
\value{
A numeric vector containing the predicted values for the response variable.
}
\description{
Generates predictions from a fitted Multivariate Adaptive Regression Splines (MARS) model.
This function uses the basis functions and model coefficients to predict the response
variable for new data. If no new data is provided, predictions are made for the data used
in fitting the \code{mars} object, equivalent to the fitted values.
}
\examples{
# Assuming 'mm' is a fitted 'mars' model
predictions <- predict(mm, newdata = some_new_data_frame)

}
\references{
Friedman, J. H. (1991). "Multivariate Adaptive Regression Splines." The Annals of Statistics,
Vol. 19, No. 1, pp. 1-67. \href{https://www.jstor.org/stable/2241837}{Link}

Additional resources on MARS model fitting, including examples, are available in the STAT 360
STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}.
}
\seealso{
\code{\link{mars}} to fit MARS models, \code{\link{residuals}} to obtain residuals.
}
