% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{split_points}
\alias{split_points}
\title{Identify Unique Split Points for MARS Model}
\usage{
split_points(xvar, Bm)
}
\arguments{
\item{xvar}{Numeric vector representing the predictor variable from which
split points are to be identified.}

\item{Bm}{Numeric vector representing the mth column of the basis function
matrix. This parameter is used to ensure that the chosen split points are
valid within the context of the current model fitting process.}
}
\value{
Returns a sorted and unique vector of valid split points derived from
\code{xvar}, excluding its maximum value to ensure meaningful partitioning of the
predictor space.
}
\description{
Determines unique and valid split points for a given predictor variable within
the context of Multivariate Adaptive Regression Splines (MARS). Split points
are critical for constructing basis functions that capture non-linear patterns
in the data. This function filters out the maximum value to avoid redundant
splits at the boundary.
}
\examples{
# Example of obtaining split points from the 'Cement' variable
# after an initial step in forward stepwise selection
fwd_result <- fwd_stepwise(y = concrete$ConcreteCompressiveStrength, x = concrete[,-1],
                           control = mars.control(Mmax = 6, d = 3, trace = FALSE))
split_pts <- split_points(concrete$Cement, fwd_result$B[, 3])

}
\references{
Friedman, J. H. (1991). "Multivariate Adaptive Regression Splines." The Annals of Statistics,
Vol. 19, No. 1, pp. 1-67. \href{https://www.jstor.org/stable/2241837}{Link to the article}

Additional insights and examples on MARS and its implementation can be found in the
STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}.
}
