% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mars.R
\name{summary.mars}
\alias{summary.mars}
\title{Summary of a MARS Model}
\usage{
\method{summary}{mars}(object)
}
\arguments{
\item{object}{A \code{mars} object produced by fitting a MARS model using the \code{mars()} function.}
}
\value{
The function does not return a value per sec but prints a detailed summary
of the MARS model to the console, akin to the summary provided for linear models (\code{lm}),
including specifics about the basis functions integrated into the model.
}
\description{
Provides a comprehensive overview of a Multivariate Adaptive Regression Splines (MARS) model,
encompassing statistical summaries, model coefficients, and information about the basis functions
utilized. This facilitates a deeper understanding of the model's structure and performance.
}
\examples{
# Assuming 'mm' is a fitted MARS model from concrete compressive strength data:
summary(mm)

}
\references{
Friedman, J. H. (1991). "Multivariate Adaptive Regression Splines." The Annals
of Statistics, Vol. 19, No. 1, pp. 1-67. Available at:
\href{https://www.jstor.org/stable/2241837}{JSTOR}.

For additional examples and documentation on MARS, visit the STAT 360 GitHub repository
STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}.
}
\seealso{
\code{\link{mars}} for fitting MARS models, and \code{\link{print.mars}} for a concise summary.
}
