% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{validate_mars.control}
\alias{validate_mars.control}
\title{Constructor for 'mars.control' Objects}
\usage{
validate_mars.control(control)
}
\arguments{
\item{control}{A list specifying the configuration for MARS model fitting. The list can include:
\itemize{
\item{\code{Mmax}}{Maximum number of basis functions to be used in the model, excluding the intercept. Should be an integer.}
\item{\code{d}}{Degree of interaction allowed in the model. Typically set to 1 (additive model) or 2 (includes interactions).}
\item{\code{trace}}{Logical flag indicating whether to trace the model fitting process. Useful for debugging or learning purposes.}
}
It's essential to provide these parameters correctly to ensure the MARS model fits as expected.}
}
\value{
Returns an object of class \code{'mars.control'}, which encapsulates the configuration for fitting a MARS model.
}
\description{
Constructs a \code{mars.control} object, specifying parameters for the fitting process of a Multivariate Adaptive Regression Splines (MARS) model. This function allows users to easily configure the key parameters that control the behavior of MARS model fitting.
}
\examples{
# Creating a mars.control object with specified parameters
mc <- new_mars.control(list(Mmax = 10, d = 3, trace = TRUE))

}
\references{
Friedman, J. H. (1991). "Multivariate Adaptive Regression Splines." The Annals of Statistics,
Vol. 19, No. 1, pp. 1-67. \href{https://www.jstor.org/stable/2241837}{DOI: https://doi.org/10.1214/aos/1176347963}

STAT 360 & STAT361 course materials: \url{https://github.com/grace84/SFUStat360Projects}.
}
\seealso{
\code{\link{mars}} for how the \code{mars.control} object is utilized in fitting MARS models.
}
